reverse string different method
This project reverses a string without using any swap or reverse. I made the reverse string function from scratch to show how each copy of the indexes work. This program takes an input string and reverses it, afterwhich it will ask if the user wishes to try to reverse another string again. it does input verification for y or n and repeats the code if input is 'y' using a do-while loop