Doggy Day Care Project (fstream practice)
This project goes over constructors and how to use fstream constructors to read files or create files.

Use the driver.cpp file to test the logic found in the other files.

Use the dummy.txt to test the open file function of the project. Or create a new txt file .

Let me know what I can improve or if you have any inputs.