Queue implementation code.
Queue has a FIFO order (first in first out). What ever is at the front of the list is what will be popped.
User enters 10 integer values that are then pushed into the queue. after queue is full then we start popping the elements starting with the element at the front index